Create teaching class
These are classes to teach the student to recognize character.

get_next_teach_sample: This is the function to get the next teach sample.
OUTPUT: Character_id: the index of character in the quiz set (0-indexed). It will return -1 if the quiz process is terminated. 

teach_judge: This is the function to update the student's learning parameters. 
INPUT: Choice: the choice that the student choose; groundtruth: the correct choice

terminated: This is the function to know if the quiz process has been terminated